refactor: remove edit context
