+++ src/components/context/ModalProviders.tsx
@@ -1,6 +1,5 @@
 import { CopyItemModalProvider } from './CopyItemModalContext';
 import { CreateShortcutModalProvider } from './CreateShortcutModalContext';
-import { EditItemModalProvider } from './EditItemModalContext';
 import { FlagItemModalProvider } from './FlagItemModalContext';
 import { LayoutContextProvider } from './LayoutContext';
 import { MoveItemModalProvider } from './MoveItemModalContext';
@@ -9,15 +8,13 @@ type Props = { children: JSX.Element };
 
 const ModalProviders = ({ children }: Props): JSX.Element => (
   <LayoutContextProvider>
-    <EditItemModalProvider>
-      <CopyItemModalProvider>
-        <MoveItemModalProvider>
-          <CreateShortcutModalProvider>
-            <FlagItemModalProvider>{children}</FlagItemModalProvider>
-          </CreateShortcutModalProvider>
-        </MoveItemModalProvider>
-      </CopyItemModalProvider>
-    </EditItemModalProvider>
+    <CopyItemModalProvider>
+      <MoveItemModalProvider>
+        <CreateShortcutModalProvider>
+          <FlagItemModalProvider>{children}</FlagItemModalProvider>
+        </CreateShortcutModalProvider>
+      </MoveItemModalProvider>
+    </CopyItemModalProvider>
   </LayoutContextProvider>
 );
